+++ apps/code-infra-dashboard/app/Providers.tsx
@@ -1,0 +1,130 @@
+'use client';
+
+import * as React from 'react';
+import Link from 'next/link';
+import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import * as colors from '@mui/material/colors';
+import CssBaseline from '@mui/material/CssBaseline';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import MuiLink from '@mui/material/Link';
+
+const theme = createTheme({
+  palette: {
+    background: { default: '#121212' },
+    primary: {
+      main: colors.amber['300'],
+    },
+    secondary: {
+      main: colors.teal['300'],
+    },
+    mode: 'dark',
+  },
+  typography: {
+    h1: { fontSize: '1.4rem' },
+    h2: { fontSize: '1.2rem' },
+    h3: { fontSize: '1.15rem' },
+    h4: { fontSize: '1.1rem' },
+    h5: { fontSize: '1.0rem' },
+    h6: { fontSize: '1.0rem' },
+  },
+  components: {
+    MuiButton: {
+      defaultProps: {
+        size: 'small',
+      },
+    },
+    MuiFilledInput: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiFormControl: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiFormHelperText: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiIconButton: {
+      defaultProps: {
+        size: 'small',
+      },
+      styleOverrides: {
+        sizeSmall: {
+          marginLeft: 4,
+          marginRight: 4,
+          padding: 12,
+        },
+      },
+    },
+    MuiInputBase: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiInputLabel: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiListItem: {
+      defaultProps: {
+        dense: true,
+      },
+    },
+    MuiOutlinedInput: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiFab: {
+      defaultProps: {
+        size: 'small',
+      },
+    },
+    MuiTable: {
+      defaultProps: {
+        size: 'small',
+      },
+    },
+    MuiTextField: {
+      defaultProps: {
+        margin: 'dense',
+      },
+    },
+    MuiToolbar: {
+      defaultProps: {
+        variant: 'dense',
+      },
+    },
+  },
+  spacing: 4,
+});
+
+const queryClient = new QueryClient();
+
+export default function Providers({ children }: { children: React.ReactNode }) {
+  return (
+    <AppRouterCacheProvider>
+      <QueryClientProvider client={queryClient}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline enableColorScheme />
+          <Container maxWidth="xl" sx={{ py: 2 }}>
+            <MuiLink component={Link} href="/" sx={{ textDecoration: 'none', color: 'inherit' }}>
+              <Typography variant="h6" component="h1" sx={{ mb: 4 }}>
+                Code infra dashboard
+              </Typography>
+            </MuiLink>
+            {children}
+          </Container>
+        </ThemeProvider>
+      </QueryClientProvider>
+    </AppRouterCacheProvider>
+  );
+}
